refactor: reorganize storage classes and types for better structure
