+++ src/storage/BasicStorage.ts
@@ -1,0 +1,6 @@
+import { AbstractStorage } from './AbstractStorage';
+
+/**
+ * BasicStorage is a basic implementation of the {@link AbstractStorage} class.
+ */
+export class BasicStorage extends AbstractStorage {}
